requirements.txt
streamlit
google-genai
folium
streamlit-folium
requests
python-dotenv